chore: dashboard api 가독성을 위해 분리
