+++ libs/api/dashboard/getInvitations.ts
@@ -1,0 +1,60 @@
+import { cookies } from 'next/headers'
+import { ApiResult } from '@/libs/types/Api'
+import { GetInvitationsResponse } from '@/libs/types/Dashboard'
+
+const BASE_URL = process.env.NEXT_PUBLIC_API_URL
+
+interface GetInvitationsParams {
+  page?: number
+  size?: number
+}
+
+export const getInvitations = async ({
+  page = 1,
+  size = 10,
+}: GetInvitationsParams): Promise<ApiResult<GetInvitationsResponse>> => {
+  try {
+    const cookieStore = await cookies()
+    const token = cookieStore.get('accessToken')?.value
+
+    const params = new URLSearchParams({
+      page: page.toString(),
+      size: size.toString(),
+    })
+
+    const response = await fetch(
+      `${BASE_URL}/invitations?${params.toString()}`,
+      {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${token}`,
+        },
+      },
+    )
+
+    const result = await response.json()
+
+    if (!response.ok) {
+      return {
+        success: false,
+        data: null,
+        error: result.message || '초대 목록을 불러오지 못했습니다.',
+      }
+    }
+
+    return {
+      success: true,
+      data: result,
+      error: null,
+    }
+  } catch (error) {
+    console.log(error)
+
+    return {
+      success: false,
+      data: null,
+      error: '네트워크 오류가 발생했습니다.',
+    }
+  }
+}
